+++ src/ETPrinter/MainWindow.xaml.cs
@@ -1,3 +1,4 @@
+using System.ComponentModel;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
@@ -10,6 +11,13 @@ public partial class MainWindow : Window
     public MainWindow()
     {
         InitializeComponent();
+        Closing += MainWindow_Closing;
+    }
+
+    private void MainWindow_Closing(object? sender, CancelEventArgs e)
+    {
+        if (!ViewModel.ConfirmDiscardChanges())
+            e.Cancel = true;
     }
 
     private MainViewModel ViewModel => (MainViewModel)DataContext;