+++ src/ETPrinter/Models/LabelProject.cs
@@ -4,7 +4,8 @@ namespace ETPrinter.Models;
 
 public class LabelProject
 {
-    public int Version { get; set; } = 2;
+    public int Version { get; set; } = 3;
+    public ProductFamily ProductFamily { get; set; } = ProductFamily.ET200SP;
     public LabelFormat Format { get; set; } = LabelFormat.HorizontalDouble;
     public LabelSettings Settings { get; set; } = new();
     public List<LabelPage> Pages { get; set; } = [];